카카오 로그인 연동 완료!

+++ GODORI/app/src/main/java/com/example/godori/data/RequestCreateUserInfo.kt
@@ -1,7 +1,7 @@
 package com.example.godori.data
 
 data class RequestCreateUserInfo(
-    val kakao_id: Int,
+    val kakao_id: Long,
     val name: String,
     val nickname: String,
     val profile_img: String,
